requirements.txt
#여기다가 쟝고에서 쓸 라이브러리 정의 하는 듯
Django==3.1.2
graphene-django==2.13.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
PyJWT==1.7.1
django-cors-headers==3.5.0
factory-boy
pytz
django-environ
gunicorn==20.0.4
djangorestframework
mysqlclient
mysql-connector-python